# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReportDetailView": {
        "additionalProperties": false,
        "description": "One report in full, envelope and content together.\n\nThe content is handed over as it was stored rather than reshaped. It is a\ndocument written at a point in time, and a reader years from now needs\nwhat was actually written, not this month's idea of how to present it.",
        "properties": {
          "content": {
            "additionalProperties": true,
            "title": "Content",
            "type": "object"
          },
          "episode_ids": {
            "items": {
              "type": "string"
            },
            "title": "Episode Ids",
            "type": "array"
          },
          "report": {
            "$ref": "#/components/schemas/ReportEnvelopeView"
          }
        },
        "required": [
          "report"
        ],
        "title": "ReportDetailView",
        "type": "object"
      },
...
      "ReportEnvelopeView": {
        "additionalProperties": false,
        "description": "One periodic report, as it appears in a list.\n\nThe envelope only. A quarter's report holds a great deal of content, and\nlisting twenty of them in full would send several megabytes to answer the\nquestion \"which reports exist\".",
        "properties": {
          "archetype_shift_detected": {
            "default": false,
            "title": "Archetype Shift Detected",
            "type": "boolean"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "episodes_analyzed": {
            "minimum": 0.0,
            "title": "Episodes Analyzed",
            "type": "integer"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "model_used": {
            "default": "",
            "title": "Model Used",
            "type": "string"
          },
          "narrative_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative Status"
          },
          "period_end": {
            "format": "date-time",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date-time",
            "title": "Period Start",
            "type": "string"
          },
          "report_id": {
            "title": "Report Id",
            "type": "string"
          },
          "report_type": {
            "title": "Report Type",
            "type": "string"
          }
        },
        "required": [
          "report_id",
          "report_type",
          "period_start",
          "period_end",
          "episodes_analyzed"
        ],
        "title": "ReportEnvelopeView",
        "type": "object"
      },
      "ReportListView": {
        "additionalProperties": false,
        "description": "A page of reports, newest first.",
        "properties": {
          "count": {
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          },
          "limit": {
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "reports": {
            "items": {
              "$ref": "#/components/schemas/ReportEnvelopeView"
            },
            "title": "Reports",
            "type": "array"
          }
        },
        "required": [
          "count",
          "limit",
          "offset"
        ],
        "title": "ReportListView",
        "type": "object"
      },
      "ReportRunRequest": {
        "additionalProperties": false,
        "description": "A request to build one report by hand.\n\nThe period is named by the day it starts. `force` writes a second report\nfor a period that already has one, which is the only way to replace a\nreport whose wording came out badly — nothing here overwrites, so both\nare kept and the newer one is what readers are shown.",
        "properties": {
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          },
          "period_start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Start"
          },
          "report_type": {
            "enum": [
              "WEEKLY",
              "MONTHLY",
              "QUARTERLY",
              "SHADOW"
            ],
            "title": "Report Type",
            "type": "string"
          }
        },
        "required": [
          "report_type"
        ],
        "title": "ReportRunRequest",
        "type": "object"
      },
      "ReportRunView": {
        "additionalProperties": false,
        "description": "What one hand-run report attempt came to.",
        "properties": {
          "duration_ms": {
            "minimum": 0.0,
            "title": "Duration Ms",
            "type": "integer"
          },
          "episodes_analyzed": {
            "minimum": 0.0,
            "title": "Episodes Analyzed",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "narrative_status": {
            "title": "Narrative Status",
            "type": "string"
          },
          "period_end": {
            "format": "date-time",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date-time",
            "title": "Period Start",
            "type": "string"
          },
          "report_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Id"
          },
          "report_type": {
            "title": "Report Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "report_type",
          "period_start",
          "period_end",
          "episodes_analyzed",
          "narrative_status",
          "duration_ms"
        ],
        "title": "ReportRunView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Read access to the knowledge graph and the history of the runs that built it, and one way in: uploading an exported conversation.",
    "title": "Lumen",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/reports": {
      "get": {
        "description": "Every report written, newest first.\n\nEnvelopes only. A quarter's report is a long document, and answering\n\"which reports exist\" with twenty of them in full would send megabytes\nto answer a question about names and dates.",
        "operationId": "list_reports_reports_get",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportListView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reports",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/reports/run": {
      "post": {
        "description": "Build one report now, without waiting for a schedule.\n\nRuns while the caller waits, because a report over a month of history is\na handful of database reads and one model call, and a background job for\nthat would need a way to report back on itself that nothing yet has.\n\nA period that already has a report is not rebuilt unless asked. That is\nwhat makes this safe to press twice.",
        "operationId": "run_report_reports_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportRunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Report",
        "tags": [
          "reports"
        ]
      }
    },
    "/reports/{report_id}": {
      "get": {
        "description": "One report in full, with the writing it drew on.\n\nThe list of episodes is what makes the report checkable. Every figure in\nit came from those and nothing else, so a reader who doubts a claim has\nsomewhere to go and look.",
        "operationId": "get_report_reports__report_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report",
        "tags": [
          "reports"
        ]
      }
    },
...
  },
...
}